libgit: support auto-clone and auto-pull of repos
Currently this just clones or pulls, as needed, a repo during the
first access to a .kbfs_autogit/tlftype/tlfname/repo directory.  It
blocks until the clone/pull is complete, or until a 10-second timeout
is hit, whichever comes first.

A future PR will deal with keeping the directory up-to-date when the
repo changes.

Issue: KBFS-2678